feat(bob): introduce v1 to v2 migration

+++ voyager/modules/client-bootstrap/base/src/main.rs
@@ -41,7 +41,7 @@ pub struct Module {
 
     pub dispute_game_factory_address: H160,
     pub dispute_game_factory_dispute_game_list_slot: U256,
-    pub fault_dispute_game_code_root_claim_index: usize,
+    pub fault_dispute_game_code_root_claim_index: u32,
 
     pub provider: DynProvider<AnyNetwork>,
 }
@@ -51,7 +51,7 @@ pub struct Module {
 pub struct Config {
     pub dispute_game_factory_address: H160,
     pub dispute_game_factory_dispute_game_list_slot: U256,
-    pub fault_dispute_game_code_root_claim_index: usize,
+    pub fault_dispute_game_code_root_claim_index: u32,
 
     /// The address of the `IBCHandler` smart contract.
     pub ibc_handler_address: H160,
